Improved Importer by making it generic. You can now specify name and description for exhibitions. Fixed some nullpointerexceptions during import.
