+++ src/main/java/ch/unibas/dmi/dbis/vrem/VREM.java
@@ -1,12 +1,12 @@
 package ch.unibas.dmi.dbis.vrem;
 
 import ch.unibas.dmi.dbis.vrem.importer.DemoImporter;
-import ch.unibas.dmi.dbis.vrem.importer.MuseumNightImporter;
+import ch.unibas.dmi.dbis.vrem.importer.ExhibitionImporter;
 import ch.unibas.dmi.dbis.vrem.server.WebServer;
 import com.github.rvesse.airline.annotations.Cli;
 
 @Cli(name = "VREM", description = "Virtual Reality Exhibition Manager", commands = {WebServer.class, DemoImporter.class,
-    MuseumNightImporter.class}, defaultCommand = WebServer.class)
+    ExhibitionImporter.class}, defaultCommand = WebServer.class)
 public class VREM {
     /**
      *
